create the OTP SDK with update event and getting refresh token from the api rel #94
