+++ package.json
@@ -25,6 +25,7 @@
   "homepage": "https://github.com/foundersandcoders/OTP-Data-Entry#readme",
   "dependencies": {
     "aws-sdk": "^2.186.0",
+    "axios": "^0.17.1",
     "body-parser": "^1.18.0",
     "compression": "^1.7.1",
     "cookie-parser": "^1.4.3",
