+++ src/otp_sdk/index.js
@@ -1,0 +1,5 @@
+module.exports.events = require('./events.js');
+
+module.exports.auth = {
+  getRefreshToken: require('./get_refresh_token.js'),
+};
